Add Jest and formatDuration tests
